+++ src/components/accounts/accountAssets/AccountIcon.vue
