+++ src/components/accounts/accountAssets/AccountIconGroup.vue
