+++ src/pages/account/AccountEditPage.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script setup>
-import AccountForm from './AccountForm.vue'
+import AccountForm from '@/components/accounts/AccountForm.vue'
 import { onMounted, reactive, ref } from 'vue'
 import { useUserStore } from '@/stores/userStore'
 import db from '/db.json'
